Refactor long method by extracting helper methods
Abstracted a long method into smaller, more manageable helper methods.
This improves code readability and maintainability by encapsulating
distinct functionality into separate methods. Each helper method
handles a specific part of the original method's logic.
