+++ src/main/java/meow/DialogBox.java
@@ -1,13 +1,16 @@
-package duke;
+package meow;
 
 import java.io.IOException;
-import java.util.Collections;
 
+import java.util.Collections;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
+
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
+
 import javafx.geometry.Pos;
+
 import javafx.scene.Node;
 import javafx.scene.control.Label;
 import javafx.scene.image.Image;
@@ -49,10 +52,25 @@ public class DialogBox extends HBox {
         dialog.getStyleClass().add("reply-label");
     }
 
+    /**
+     * Creates and returns a DialogBox for the user's input.
+     *
+     * @param text The user's input text.
+     * @param img The user's profile image.
+     * @return A DialogBox representing the user's dialog.
+     */
     public static DialogBox getUserDialog(String text, Image img) {
         return new DialogBox(text, img);
     }
 
+    /**
+     * Creates and returns a DialogBox for Meow's response.
+     * The dialog is flipped so that it is aligned as a response.
+     *
+     * @param text The response text from Meow.
+     * @param img The profile image of Meow.
+     * @return A DialogBox representing Meow's dialog, flipped for response alignment.
+     */
     public static DialogBox getMeowDialog(String text, Image img) {
         var db = new DialogBox(text, img);
         db.flip();
