+++ src/main/java/meow/Launcher.java
@@ -1,0 +1,20 @@
+package meow;
+
+import javafx.application.Application;
+
+/**
+ * The Launcher class is the entry point to the application.
+ * It launches the JavaFX application
+ */
+public class Launcher {
+
+    /**
+     * Standard main (psvm) method
+     *
+     * @param args the command line arguments passed to the application
+     */
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
+
