+++ src/main/java/meow/Meow.java
@@ -1,6 +1,7 @@
-package duke;
+package meow;
 
 import java.io.IOException;
+
 import java.util.ArrayList;
 
 /**
