+++ src/main/java/meow/Task.java
@@ -1,4 +1,4 @@
-package duke;
+package meow;
 
 import java.io.Serializable;
