+++ src/main/java/meow/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package meow;
 
 import java.util.Scanner;
