+++ src/test/java/duke/ParserTest.java
@@ -1,5 +1,6 @@
 package duke;
 
+import meow.Parser;
 import org.junit.jupiter.api.Test;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
